http: deprecate undocumented internal modules
