+++ lib/http.js
@@ -25,18 +25,18 @@ const {
   ObjectDefineProperty,
 } = primordials;
 
-const httpAgent = require('_http_agent');
-const { ClientRequest } = require('_http_client');
-const { methods } = require('_http_common');
-const { IncomingMessage } = require('_http_incoming');
-const { OutgoingMessage } = require('_http_outgoing');
+const httpAgent = require('internal/http/agent');
+const { ClientRequest } = require('internal/http/client');
+const { methods } = require('internal/http/common');
+const { IncomingMessage } = require('internal/http/incoming');
+const { OutgoingMessage } = require('internal/http/outgoing');
 const { validateHeaderName, validateHeaderValue } = OutgoingMessage;
 const {
   _connectionListener,
   STATUS_CODES,
   Server,
   ServerResponse
-} = require('_http_server');
+} = require('internal/http/server');
 let maxHeaderSize;
 
 function createServer(opts, requestListener) {
