+++ lib/https.js
@@ -30,17 +30,17 @@ require('internal/util').assertCrypto();
 
 const tls = require('tls');
 const url = require('url');
-const { Agent: HttpAgent } = require('_http_agent');
+const { Agent: HttpAgent } = require('internal/http/agent');
 const {
   Server: HttpServer,
   _connectionListener,
   kServerResponse
-} = require('_http_server');
-const { ClientRequest } = require('_http_client');
+} = require('internal/http/server');
+const { ClientRequest } = require('internal/http/client');
 const debug = require('internal/util/debuglog').debuglog('https');
 const { URL, urlToOptions, searchParamsSymbol } = require('internal/url');
 const { IncomingMessage, ServerResponse } = require('http');
-const { kIncomingMessage } = require('_http_common');
+const { kIncomingMessage } = require('internal/http/common');
 
 function Server(opts, requestListener) {
   if (!(this instanceof Server)) return new Server(opts, requestListener);
