+++ lib/internal/http2/core.js
@@ -35,8 +35,8 @@ const tls = require('tls');
 const { URL } = require('url');
 const { setImmediate } = require('timers');
 
-const { kIncomingMessage } = require('_http_common');
-const { kServerResponse } = require('_http_server');
+const { kIncomingMessage } = require('internal/http/common');
+const { kServerResponse } = require('internal/http/server');
 const JSStreamSocket = require('internal/js_stream_socket');
 
 const {
