+++ node.gyp
@@ -141,6 +141,12 @@
       'lib/internal/fs/utils.js',
       'lib/internal/fs/watchers.js',
       'lib/internal/http.js',
+      'lib/internal/http/agent.js',
+      'lib/internal/http/client.js',
+      'lib/internal/http/common.js',
+      'lib/internal/http/incoming.js',
+      'lib/internal/http/outgoing.js',
+      'lib/internal/http/server.js',
       'lib/internal/heap_utils.js',
       'lib/internal/histogram.js',
       'lib/internal/idna.js',
