+++ test/parallel/test-warn-http-agent.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const common = require('../common');
+
+// _http_agent is deprecated.
+
+common.expectWarning('DeprecationWarning',
+                     'The _http_agent module is deprecated.', 'DEPXXXX');
+
+require('_http_agent');
