+++ test/parallel/test-warn-http-outgoing.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const common = require('../common');
+
+// _http_outgoing is deprecated.
+
+common.expectWarning('DeprecationWarning',
+                     'The _http_outgoing module is deprecated.', 'DEPXXXX');
+
+require('_http_outgoing');
